@@ -2,6 +2,7 @@
     "name": "slam/php-validatore-fattura-elettronica",
     "description": "Validatore fattura elettronica",
     "license": "MIT",
+    "type": "library",
     "authors": [
         {
             "name": "Filippo Tessarotto",
@@ -10,16 +11,16 @@
         }
     ],
     "require": {
-        "php": ">=7.2.0",
+        "php": "~8.1.0 || ~8.2.0",
         "ext-dom": "*",
         "ext-libxml": "*"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^v3.8.0",
-        "phpstan/phpstan": "^1.5.4",
-        "phpstan/phpstan-phpunit": "^1.1.0",
-        "phpstan/phpstan-strict-rules": "^1.1.0",
-        "phpunit/phpunit": "^8.5.26 || ^9.5.20"
+        "phpstan/phpstan": "^1.10.1",
+        "phpstan/phpstan-phpunit": "^1.3.7",
+        "phpstan/phpstan-strict-rules": "^1.5.0",
+        "phpunit/phpunit": "^10.0.11",
+        "slam/php-cs-fixer-extensions": "^3.4.0"
     },
     "autoload": {
         "psr-4": {
